rename header stage to validate_header
