+++ crates/amaru/src/stages/consensus/mod.rs
@@ -14,4 +14,4 @@
 
 pub mod chain_forward;
 pub mod fetch;
-pub mod header;
+pub mod validate_header;
